requirements.txt
aiosqlite==0.17.0
annotated-types==0.6.0
anyio==4.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
diffusers==0.25.1
exceptiongroup==1.2.0
fastapi==0.109.0
filelock==3.13.1
fsspec==2023.12.2
h11==0.14.0
huggingface-hub==0.20.3
idna==3.6
importlib-metadata==7.0.1
iso8601==1.1.0
Jinja2==3.1.3
MarkupSafe==2.1.4
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
packaging==23.2
pillow==10.2.0
pydantic==2.5.3
pydantic_core==2.14.6
pypika-tortoise==0.1.6
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.1
scipy==1.12.0
sniffio==1.3.0
starlette==0.35.1
sympy==1.12
tokenizers==0.15.1
torch==2.1.2
tortoise-orm==0.20.0
tqdm==4.66.1
transformers==4.37.0
typing_extensions==4.9.0
urllib3==2.1.0
uvicorn==0.26.0
zipp==3.17.0
python-multipart==0.0.6
